+++ server/impl/src/main/java/com/walmartlabs/concord/server/metrics/GaugeProvider.java
@@ -1,0 +1,10 @@
+package com.walmartlabs.concord.server.metrics;
+
+import com.codahale.metrics.Gauge;
+
+public interface GaugeProvider<T> {
+
+    String name();
+
+    Gauge<T> gauge();
+}