+++ server/impl/src/main/java/com/walmartlabs/concord/server/metrics/MetricRegistryProvider.java
@@ -25,13 +25,22 @@ import io.prometheus.client.CollectorRegistry;
 import io.prometheus.client.dropwizard.DropwizardExports;
 import io.prometheus.client.hotspot.DefaultExports;
 
+import javax.inject.Inject;
 import javax.inject.Named;
 import javax.inject.Provider;
 import javax.inject.Singleton;
+import java.util.Set;
 
 @Named
 @Singleton
 public class MetricRegistryProvider implements Provider<MetricRegistry> {
+
+    private final Set<GaugeProvider<?>> gauges;
+
+    @Inject
+    public MetricRegistryProvider(Set<GaugeProvider<?>> gauges) {
+        this.gauges = gauges;
+    }
 
     @Override
     public MetricRegistry get() {
@@ -42,6 +51,9 @@ public class MetricRegistryProvider implements Provider<MetricRegistry> {
 
         // initialize standard prometheus exports (hotspot, memory, etc)
         DefaultExports.initialize();
+
+        // register gauges
+        gauges.forEach(g -> registry.register(g.name(), g.gauge()));
 
         return registry;
     }